v1.2.1 完善 cli 配置
