+++ bbot/cli/config.py
@@ -1,7 +1,6 @@
+import sys
 import click
 import httpx
-
-from typing import Optional
 
 from yarl import URL
 from pathlib import Path
@@ -15,8 +14,9 @@ from noneprompt import (
     CheckboxPrompt,
 )
 
+from ..model.config import _BotConfig
 from ..core.announcement import BBOT_ASCII_LOGO
-from ..core.bot_config import BotConfig
+
 
 data = {}
 
@@ -28,47 +28,22 @@ class CliConfig:
         "Bilibili": {},
         "Webui": {},
     }
-    """{
-        "Mirai": {
-            "mirai_host": "https://localhost:8080",
-            "verify_key": "xxxxxxxxx",
-            "account": 123456789,
-        },
-        "Debug": {"enable": false, "groups": [123456789]},
-        "Bilibili": {
-            "mobile_style": true,
-            "use_login": false,
-            "username": null,
-            "password": null,
-            "concurrency": 5,
-            "dynamic_font": "https://fonts.bbot?name=https://cdn.jsdelivr.net/gh/irozhi/HarmonyOS-Sans/HarmonyOS_Sans_SC/HarmonyOS_Sans_SC_Medium.woff2",
-            "dynamic_font_source": "remote",
-        },
-        "Event": {"mute": true, "permchange": true, "push": true, "subscribe": true},
-        "Webui": {"webui_host": "0.0.0.0", "webui_port": 6080, "webui_enable": true},
-        "log_level": "INFO",
-        "name": "BBot",
-        "access_control": true,
-        "master": 123456789,
-        "admins": [123456789],
-        "max_subsubscribe": 4,
-    }"""
     __session: str = ""
 
     def __init__(self) -> None:
-        self.mirai_mirai_host()  # pass
-        self.mirai_verify_key()  # pass
-        self.debug()  # pass
-        self.bilibili_mobile_style()  # pass
-        self.bilibili_concurrent()  # pass
-        self.event()  # pass
-        self.webui()  # pass
-        self.log_level()  # pass
-        self.name()  # pass
-        self.access_control()  # pass
-        self.master()  # pass
-        self.admins()  #
-        self.max_subscriptions()  #
+        self.mirai_mirai_host()
+        self.mirai_verify_key()
+        self.debug()
+        self.bilibili_mobile_style()
+        self.bilibili_concurrent()
+        self.event()
+        self.webui()
+        self.log_level()
+        self.name()
+        self.access_control()
+        self.master()
+        self.admins()
+        self.max_subscriptions()
 
         if data.get("session", None):
             httpx.post(
@@ -99,7 +74,7 @@ class CliConfig:
             if self.is_qq(mirai_account):
                 self.config["Mirai"]["account"] = int(mirai_account)
                 return
-            click.secho("输入的 QQ 号不合法！", fg="red", bold=True)
+            click.secho("输入的 QQ 号不合法！", fg="bright_red", bold=True)
 
     # ----- Mirai -----
     def mirai_mirai_host(self):
@@ -111,25 +86,27 @@ class CliConfig:
                 if not URL(mirai_host).is_absolute():
                     raise ValueError("输入的地址不合法！")
             except Exception:
-                click.secho("输入的地址不合法！", fg="red", bold=True)
+                click.secho("输入的地址不合法！", fg="bright_red", bold=True)
                 continue
 
             try:
                 if httpx.get(f"{mirai_host}/about").json()["data"]["version"] < "2.6.1":
                     click.secho(
                         "Mirai HTTP API 版本低于 2.6.1，可能会导致部分功能无法使用！请升级至最新版",
-                        fg="red",
+                        fg="bright_red",
                         bold=True,
                     )
                 self.config["Mirai"]["mirai_host"] = mirai_host
                 return
             except httpx.HTTPError:
-                click.secho("无法连接到 Mirai HTTP API，请检查地址是否正确！", fg="red", bold=True)
+                click.secho("无法连接到 Mirai HTTP API，请检查地址是否正确！", fg="bright_red", bold=True)
                 continue
 
     def mirai_verify_key(self):
         while True:
-            mirai_key: str = InputPrompt("请输入 Mirai HTTP API 的 verifyKey: ").prompt()
+            mirai_key: str = InputPrompt(
+                "请输入 Mirai HTTP API 的 verifyKey: ", is_password=True
+            ).prompt()
             try:
                 # check verifyKey
                 verify = httpx.post(
@@ -137,7 +114,7 @@ class CliConfig:
                     json={"verifyKey": mirai_key},
                 ).json()
                 if verify["code"] != 0:
-                    click.secho("Mirai HTTP API 的 verifyKey 错误！", fg="red", bold=True)
+                    click.secho("Mirai HTTP API 的 verifyKey 错误！", fg="bright_red", bold=True)
                     continue
                 self.__session = verify["session"]
 
@@ -160,19 +137,21 @@ class CliConfig:
                     ):
                         break
                     click.secho(
-                        f"Mirai HTTP API 验证错误（可能是QQ号填写错误）：{bind['msg']}！", fg="red", bold=True
+                        f"Mirai HTTP API 验证错误（可能是QQ号填写错误）：{bind['msg']}！",
+                        fg="bright_red",
+                        bold=True,
                     )
 
                 # all clear
-                click.secho("Mirai HTTP API 验证成功！", fg="green", bold=True)
+                click.secho("Mirai HTTP API 验证成功！", fg="bright_green", bold=True)
                 self.config["Mirai"]["verify_key"] = mirai_key
                 return
 
             except httpx.HTTPError:
-                click.secho("无法连接到 Mirai HTTP API，请检查地址是否正确！", fg="red", bold=True)
+                click.secho("无法连接到 Mirai HTTP API，请检查地址是否正确！", fg="bright_red", bold=True)
                 self.mirai_mirai_host()
             except JSONDecodeError:
-                click.secho("输入的地址不为 Mirai HTTP API 的地址！", fg="red", bold=True)
+                click.secho("输入的地址不为 Mirai HTTP API 的地址！", fg="bright_red", bold=True)
                 self.mirai_mirai_host()
 
     def debug(self):
@@ -195,14 +174,14 @@ class CliConfig:
             debug_group = InputPrompt("请输入调试群的群号（输入 n 停止）: ").prompt()
             if debug_group.lower() == "n":
                 if not self.config["Debug"]["groups"]:
-                    click.secho("请至少输入一个群号！", fg="red", bold=True)
+                    click.secho("请至少输入一个群号！", fg="bright_red", bold=True)
                     continue
                 return
             if not debug_group.isdigit():
-                click.secho("输入的群号不合法！", fg="red", bold=True)
+                click.secho("输入的群号不合法！", fg="bright_red", bold=True)
                 continue
             if int(debug_group) in self.config["Debug"]["groups"]:
-                click.secho("该群已在调试群列表中，请重新输入！", fg="red", bold=True)
+                click.secho("该群已在调试群列表中，请重新输入！", fg="bright_red", bold=True)
                 continue
             debug_group = int(debug_group)
             group_list = httpx.get(
@@ -215,7 +194,7 @@ class CliConfig:
                     self.config["Debug"]["groups"].append(debug_group)
                     break
             else:
-                click.secho("Bot 未加入该群，请重新输入！", fg="red", bold=True)
+                click.secho("Bot 未加入该群，请重新输入！", fg="bright_red", bold=True)
 
     def bilibili_mobile_style(self):
         mobile_style = ListPrompt(
@@ -232,7 +211,7 @@ class CliConfig:
             if concurrent.isdigit() and 50 >= int(concurrent) > 0:
                 self.config["Bilibili"]["concurrency"] = int(concurrent)
                 return
-            click.secho("输入的并发数不合法！请输入 1 - 50 之间的整数", fg="red", bold=True)
+            click.secho("输入的并发数不合法！请输入 1 - 50 之间的整数", fg="bright_red", bold=True)
 
     def event(self):
         event_map = {"禁言": "mute", "权限变更": "permission", "动态推送": "push", "UP 订阅": "subscribe"}
@@ -270,7 +249,7 @@ class CliConfig:
                 if not url.is_absolute():
                     raise ValueError("输入的地址不合法！")
             except Exception:
-                click.secho("输入的地址不合法！", fg="red", bold=True)
+                click.secho("输入的地址不合法！", fg="bright_red", bold=True)
                 continue
             self.config["Webui"] = {
                 "webui_host": url.host,
@@ -282,51 +261,51 @@ class CliConfig:
     # ----- Other -----
     def master(self):
         while True:
-            master = InputPrompt("请输入你的 QQ 号(作为最高管理员账号): ").prompt()
+            master = InputPrompt("请输入你的 QQ 号（作为最高管理员账号）: ").prompt()
             if not self.is_qq(master):
-                click.secho("输入的 QQ 号不合法！", fg="red", bold=True)
+                click.secho("输入的 QQ 号不合法！", fg="bright_red", bold=True)
                 continue
 
             if self.verify_friend(master):
                 self.config["master"] = int(master)
                 return
             else:
-                click.secho("你输入的 QQ 号不是 Bot 的好友！", fg="red", bold=True)
+                click.secho("你输入的 QQ 号不是 Bot 的好友！", fg="bright_red", bold=True)
 
     def admins(self):
         self.config["admins"] = [self.config["master"]]
         while True:
             admin = InputPrompt("请输入其他管理员的 QQ 号（输入 n 停止）: ").prompt()
-            if admin.lower() == "n":
+            if admin.lower() == "n" or admin == "":
                 return
             if not self.is_qq(admin):
-                click.secho("输入的 QQ 号不合法！", fg="red", bold=True)
+                click.secho("输入的 QQ 号不合法！", fg="bright_red", bold=True)
                 continue
             if int(admin) in self.config["admins"]:
-                click.secho("该 QQ 号已经在管理员列表中了！", fg="red", bold=True)
+                click.secho("该 QQ 号已经在管理员列表中了！", fg="bright_red", bold=True)
                 continue
             if not self.verify_friend(admin):
-                click.secho("该 QQ 号不是你的好友！", fg="red", bold=True)
+                click.secho("该 QQ 号不是 Bot 的好友！", fg="bright_red", bold=True)
                 continue
 
             self.config["admins"].append(int(admin))
 
     def log_level(self):
-        self.config["log_level"] = ListPrompt(
-            "请选择日志等级",
-            [
-                Choice("TRACE"),
-                Choice("DEBUG"),
-                Choice("INFO"),
-                Choice("SUCCESS"),
-                Choice("WARNING"),
-                Choice("ERROR"),
-                Choice("CRITICAL"),
-            ],
-            allow_filter=False,
-            default_select=2,
-            annotation="使用键盘的 ↑ 和 ↓ 来选择, 按回车确认",
-        ).prompt().name
+        self.config["log_level"] = (
+            ListPrompt(
+                "请选择日志等级",
+                [
+                    Choice("DEBUG"),
+                    Choice("INFO"),
+                    Choice("WARNING"),
+                ],
+                allow_filter=False,
+                default_select=1,
+                annotation="使用键盘的 ↑ 和 ↓ 来选择, 按回车确认",
+            )
+            .prompt()
+            .name
+        )
 
     def name(self):
         while True:
@@ -334,7 +313,7 @@ class CliConfig:
             if 0 < len(bot_name) <= 16:
                 self.config["name"] = bot_name
                 return
-            click.secho("输入的昵称不合法（请输入 1 - 16 位字符）", fg="red", bold=True)
+            click.secho("输入的昵称不合法（请输入 1 - 16 位字符）", fg="bright_red", bold=True)
 
     def max_subscriptions(self):
         while True:
@@ -342,7 +321,7 @@ class CliConfig:
             if max_subscriptions.isdigit() and 0 < int(max_subscriptions) <= 100:
                 self.config["max_subsubscribe"] = int(max_subscriptions)
                 return
-            click.secho("输入的订阅数不合法！请输入 1 - 100 之间的整数", fg="red", bold=True)
+            click.secho("输入的订阅数不合法！请输入 1 - 100 之间的整数", fg="bright_red", bold=True)
 
     def access_control(self):
         access_control = ListPrompt(
@@ -354,48 +333,37 @@ class CliConfig:
         self.config["access_control"] = access_control.name == "是（开启）"
 
 
-def click_config():
+def click_config(ignore_exist=False):
     click.secho(BBOT_ASCII_LOGO, fg="bright_blue", bold=True)
-    click.secho("\n欢迎使用 BBot 配置向导！\n", fg="green", bold=True)
-    # click.secho("配置向导仍在开发，暂不可用。", fg="red", bold=True)
-    # return
-    click.secho("请按照提示输入相关信息。\n", fg="yellow")
+    click.secho("\n欢迎使用 BBot 配置向导！\n", fg="bright_green", bold=True)
+    click.secho("请按照提示输入相关信息后，按回车确认\n", fg="bright_yellow")
 
     try:
-        if list(Path(".").iterdir()):
-            click.secho("当前目录不为空，可能会导致未知的错误。", fg="yellow", bold=True)
-            click.secho("请确保当前目录没有 Bot 之外的文件。", fg="yellow", bold=True)
-
-            if not ConfirmPrompt("是否仍要继续配置流程？", default_choice=False).prompt():
-                return
-
-        config_method = ListPrompt(
-            "请选择你要使用的配置方式：",
-            [
-                # Choice("使用 Web 配置向导", "web"),
-                Choice("使用命令行配置向导", "cli"),
-                Choice("手动修改配置文件", "file"),
-            ],
-            annotation="使用键盘的 ↑ 和 ↓ 来选择, 按回车确认",
-        ).prompt()
-
-        if config_method.data == "cli":
-            BotConfig.parse_obj(CliConfig().config).save()
-        # elif config_method.data == "web":
-        #     click.secho("Web 配置向导暂不可用。", fg="red", bold=True)
-        else:
-            BotConfig._create_file()
-            click.secho("配置文件已创建（data/bot_config.yaml）。请手动打开后进行编辑。", fg="green", bold=True)
-            while True:
-                if ConfirmPrompt("是否完成了配置？", default_choice=False).prompt():
-                    config_verify = BotConfig.verify()
-                    if not config_verify:
-                        break
-                    click.secho("配置文件有误，请检查后重试。错误信息如下：", fg="red", bold=True)
-                    click.secho(config_verify, fg="red", bold=True)
-
-        click.secho("恭喜你，配置已完成！\n", fg="green", bold=True)
-        click.secho("现在，你可以使用 bbot run 命令运行 BBot 了。", fg="blue", bold=True)
-
+        _config(ignore_exist)
+        sys.exit(0)
     except CancelledError:
-        click.secho("配置向导已取消。", fg="yellow", bold=True)
+        click.secho("配置向导已取消。", fg="bright_yellow", bold=True)
+        sys.exit(0)
+
+
+def _config(ignore_exist):
+    if list(Path(".").iterdir()) and not ignore_exist:
+        click.secho("当前目录不为空，可能会导致未知的错误。", fg="bright_yellow", bold=True)
+        click.secho("请确保当前目录没有 Bot 之外的文件。", fg="bright_yellow", bold=True)
+
+        if not ConfirmPrompt("是否仍要继续配置流程？", default_choice=False).prompt():
+            sys.exit(0)
+        if Path("data", "bot_config.yaml").exists():
+            click.secho("检测到已存在的配置文件，将覆盖原配置。", fg="bright_yellow", bold=True)
+            if not ConfirmPrompt("是否仍要继续配置流程？", default_choice=False).prompt():
+                sys.exit(0)
+
+    if not Path("data").exists():
+        click.secho("data 目录不存在，已自动创建", fg="bright_yellow")
+        Path("data").mkdir(parents=True, exist_ok=True)
+
+    click.secho("正在进入配置流程……", fg="bright_green", bold=True)
+    _BotConfig.parse_obj(CliConfig().config).save()
+
+    click.secho("恭喜你，配置已完成！", fg="bright_green", bold=True)
+    click.secho("现在，你可以使用 bbot run 命令运行 BBot 了。", fg="bright_blue", bold=True)
