+++ bbot/cli/run.py
@@ -1,4 +1,5 @@
 import contextlib
+
 from loguru import logger
 from sentry_sdk import init as sentry_sdk_init
