+++ bbot/static/bot_config.exp.yaml
@@ -16,7 +16,7 @@ Bilibili:
   username:          # 若登录 BiliBili 账号, 则 **需要** 填入账号
   password:          # 若登录 BiliBili 账号, 则 **需要** 填入密码
   concurrency: 5     # 未登录时发送 gRPC 请求的并发数量(1 ~ 50)
-  dynamic_font: "https://fonts.bbot?name=https://cdn.jsdelivr.net/gh/irozhi/HarmonyOS-Sans/HarmonyOS_Sans_SC/HarmonyOS_Sans_SC_Medium.woff2" # 自定义动态字体
+  dynamic_font: "https://cdn.jsdelivr.net/gh/irozhi/HarmonyOS-Sans/HarmonyOS_Sans_SC/HarmonyOS_Sans_SC_Medium.woff2" # 自定义动态字体
   dynamic_font_source: "remote" # 自定义动态字体来源, 可选值: "local", "remote", "system", 为 "local" 时请将字体文件放在 data/font 目录下
 Event:
   mute: true         # 是否向管理员发送被禁言的事件提醒。
