+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "bbot"
-version = "1.2.0-beta5"
+version = "1.2.1"
 description = "一个用于 QQ 群内高效推送哔哩哔哩 UP 动态及直播的机器人"
 authors = ["djkcyl <cyl@cyllive.cn>"]
 license = "AGPL3.0"
@@ -22,7 +22,7 @@ bilireq = "^0.2.3"
 psutil = "^5.9.4"
 graiax-playwright = "^0.2.1"
 uvicorn = "^0.19.0"
-sentry-sdk = "^1.12.1"
+sentry-sdk = "^1.13.0"
 python-jose = {extras = ["cryptography"], version = "^3.3.0"}
 passlib = {extras = ["bcrypt"], version = "^1.7.4"}
 python-multipart = "^0.0.5"
