+++ requirements.txt
@@ -28,16 +28,18 @@ greenlet==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
 grpcio==1.51.1 ; python_version >= "3.9" and python_version < "4.0"
 h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
 httpcore==0.16.3 ; python_version >= "3.9" and python_version < "4.0"
-httpx==0.23.2 ; python_version >= "3.9" and python_version < "4.0"
+httpx==0.23.3 ; python_version >= "3.9" and python_version < "4.0"
 idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
 importlib-metadata==6.0.0 ; python_version >= "3.9" and python_version < "4.0"
-launart==0.6.2 ; python_version >= "3.9" and python_version < "4.0"
+launart==0.6.3 ; python_version >= "3.9" and python_version < "4.0"
 loguru==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
 multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
+noneprompt==0.1.7 ; python_version >= "3.9" and python_version < "4.0"
 passlib[bcrypt]==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
 peewee==3.15.4 ; python_version >= "3.9" and python_version < "4.0"
 pillow==9.4.0 ; python_version >= "3.9" and python_version < "4.0"
-playwright==1.29.0 ; python_version >= "3.9" and python_version < "4.0"
+playwright==1.29.1 ; python_version >= "3.9" and python_version < "4.0"
+prompt-toolkit==3.0.36 ; python_version >= "3.9" and python_version < "4.0"
 protobuf==4.21.12 ; python_version >= "3.9" and python_version < "4.0"
 psutil==5.9.4 ; python_version >= "3.9" and python_version < "4.0"
 pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
@@ -51,17 +53,18 @@ python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
 pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
 qrcode==7.3.1 ; python_version >= "3.9" and python_version < "4.0"
 rfc3986[idna2008]==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
-rich==13.0.0 ; python_version >= "3.9" and python_version < "4.0"
+rich==13.1.0 ; python_version >= "3.9" and python_version < "4.0"
 richuru==0.1.1 ; python_version >= "3.9" and python_version < "4.0"
 rsa==4.9 ; python_version >= "3.9" and python_version < "4"
-sentry-sdk==1.12.1 ; python_version >= "3.9" and python_version < "4.0"
+sentry-sdk==1.13.0 ; python_version >= "3.9" and python_version < "4.0"
 six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
 sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
 starlette==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
 statv==0.3.2 ; python_version >= "3.9" and python_version < "4.0"
 typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
-urllib3==1.26.13 ; python_version >= "3.9" and python_version < "4.0"
+urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4.0"
 uvicorn==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
+wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
 websockets==10.4 ; python_version >= "3.9" and python_version < "4.0"
 win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
 yarl==1.8.2 ; python_version >= "3.9" and python_version < "4.0"
